+++ basic/src/main/java/com/dede/basic/Utils.kt
@@ -3,11 +3,13 @@
 
 package com.dede.basic
 
+import android.animation.ValueAnimator
 import android.app.Activity
 import android.content.Context
 import android.content.Intent
 import android.graphics.Color
 import android.os.Build
+import android.provider.Settings
 import android.view.WindowManager
 import androidx.activity.ComponentActivity
 import androidx.activity.enableEdgeToEdge
@@ -16,6 +18,33 @@ import androidx.core.view.WindowInsetsCompat
 import androidx.core.view.WindowInsetsControllerCompat
 
 object Utils {
+
+    fun getDevOptsAnimatorDurationScaleIntent(): Intent {
+        val intent = Intent(Settings.ACTION_APPLICATION_DEVELOPMENT_SETTINGS)
+        intent.putExtra(":settings:fragment_args_key", "animator_duration_scale")
+        intent.addFlags(Intent.FLAG_ACTIVITY_NEW_TASK)
+        return intent
+    }
+
+    fun areAnimatorEnabled(context: Context): Boolean {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.O) {
+            ValueAnimator.areAnimatorsEnabled()
+        } else {
+            getAnimatorDurationScale(context) != 0f
+        }
+    }
+
+    fun getAnimatorDurationScale(context: Context): Float {
+        return if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+            ValueAnimator.getDurationScale()
+        } else {
+            Settings.Global.getFloat(
+                context.contentResolver,
+                Settings.Global.ANIMATOR_DURATION_SCALE,
+                1f
+            )
+        }
+    }
 
     fun getLaunchIntent(context: Context): Intent? {
         val pm = context.packageManager