+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### v3.2.0
 
+- Add 'Animator duration scale' alert
 - Project modularization
 - Upgrade project dependencies